Define CIString To Be A CanonicalFullCaseFoldedString
This allows us to not deprecate the type.

There is still an unfortunate issue here. Full case folding can change the length (in terms of characters) of the `String`. This makes `CIString.length` not safe to use, as it is not clear which length it now refers to. This commit deprecates that method because of this. However, the `glob` based `unapplySeq` relies both on `length` and `Char` based case insensitivity. This means that it is no longer correct.

That said, by relying on `Char` based case insensitivity, it was probably not quite correct to begin with.